backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis
redis==5.0.1
hiredis==2.2.3

# AI/ML
openai==1.3.7
google-generativeai==0.3.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2

# Payment gateways
stripe==7.8.0
paypalrestsdk==1.13.3
razorpay==1.4.1

# Network and hardware integration
paramiko==3.4.0
netmiko==4.3.0
pysnmp==4.4.12

# Security
cryptography==41.0.8
python-multipart==0.0.6

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Utilities
python-decouple==3.8
celery==5.3.4
flower==2.0.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# CORS
python-cors==1.7.0